move redteam types
