+++ src/redteam/types.ts
