+++ test/redteam/commands/generate.test.ts
@@ -3,7 +3,7 @@ import * as configModule from '../../../src/config';
 import logger from '../../../src/logger';
 import { synthesize } from '../../../src/redteam';
 import { doGenerateRedteam } from '../../../src/redteam/commands/generate';
-import type { RedteamCliGenerateOptions } from '../../../src/types/redteam';
+import type { RedteamCliGenerateOptions } from '../../../src/redteam/types';
 import { writePromptfooConfig } from '../../../src/util/config';
 
 jest.mock('fs');
